Improve precision of DisVerTri
If the point was very close from an edge the test of barycentric
coordinates was very unprecise because it was adding small and large
values. Heavy testing show that this test is unneeded. Checking on which
side of each edge is the point is enough. It's also a bit faster.